CouponManagement/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
dotenv
psycopg2-binary